// Public/shared/src/models/Product.Model.ts
export interface IProductModel {
    Id: number;
    Name?: string;
    ProductId: number;
    EncryptedProductId: string;
    EncryptedId: string;
    ProductTypeId: number;
    ParentGroupedProductId?: number;
    Visibility: number;
    VisibleIndividually: boolean;
    Condition: string;
    ProductName: string;
    ShortDescription: string;
    FullDescription: string;
    ProductTemplateId: number;
    ShowOnHomePage: boolean;
    HomePageDisplayOrder: number;
    MetaKeywords: string;
    MetaDescription: string;
    MetaTitle: string;
    ManufacturerPartNumber: string;
    Gtin: string;
    RequireOtherProducts: boolean;
    RequiredProductIds: string;
    AutomaticallyAddRequiredProducts: boolean;
    IsRecurring: boolean;
    IsShipEnabled: boolean;
    IsFreeShipping: boolean;
    AdditionalShippingCharge: number;
    TaxCategoryId: number;
    DisplayStockAvailability: boolean;
    DisplayStockQuantity: boolean;
    MinStockQuantity: number;
    LowStockActivityId: number;
    NotifyAdminForQuantityBelow: number;
    BackorderModeId: number;
    DisableBuyButton: boolean;
    DisableWishlistButton: boolean;
    AvailableForPreOrder: boolean;
    CallForPrice: boolean;
    Price: number;
    FinalTotalPrice: any;
    OldPrice: number;
    ComparePriceLabelId?: number;
    ProductCost: number;
    SpecialPrice?: number;
    SpecialPriceStartDateTimeUtc?: Date;
    SpecialPriceEndDateTimeUtc?: Date;
    CustomerEntersPrice: boolean;
    MinimumCustomerEnteredPrice: number;
    MaximumCustomerEnteredPrice: number;
    HasTierPrices: boolean;
    LowestAttributeCombinationPrice: number;
    AttributeChoiceBehaviour: number;
    Weight: number;
    Length: number;
    Width: number;
    Height: number;
    DisplayOrder: number;
    Published: boolean;
    Deleted: boolean;
    IsSystemProduct: boolean;
    SystemName: string;
    CreatedOnUtc: Date;
    UpdatedOnUtc: Date;
    DeliveryTimeId?: number;
    QuantityUnitId?: number;
    CustomsTariffNumber: string;
    CountryOfOriginId?: number;
    BasePriceEnabled: boolean;
    BasePriceMeasureUnit: string;
    BasePriceAmount: number;
    BasePriceBaseAmount: number;
    BundleTitleText: string;
    BundlePerItemShipping: boolean;
    BundlePerItemPricing: boolean;
    BundlePerItemShoppingCart: boolean;
    MainPictureId?: number;
    HasPreviewPicture: boolean;
    HasDiscountsApplied: boolean;
    ProductAttributeId: number;
    ProductAttributes: any;
    ProductAttributeOption:any;
    SpecificationAttributeName: string;
    SpecificationAttributeOptionName: string;
    DiscountName: string;
    TotalPrice:number;
    PriceAdjustment: string;
    RequiresCouponCode: boolean;
    CouponCode:any;
    IsValid:boolean;
    ProductVariantAttributeValue:any;
    FormattedMediaFileName:any;
    IsTaxExempt: boolean;
    AttributesXml: string;
    DiscountPricePercentage:any;
    IsVariant?: boolean,
    AssignedMedia?:any,
    ImagePath?: string,
    ColorImages?: {
    ColorID: number;
    ColorName: string;
    ImagePath: string;
  }[];
}